requirements.txt
asgiref==3.7.2
Django==3.2.25
django-allauth==0.54.0
djangorestframework==3.13.1
djangorestframework_simplejwt==5.2.2
drf-yasg==1.20.0
inflection==0.5.1
loadenv==0.1.1
packaging==23.2
psycopg2-binary==2.9.6
PyJWT==2.6.0
python-decouple==3.6
python-dotenv==0.21.0
pytz==2023.3
PyYAML==5.4.1
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==3.0.1

